chore(test): rename test file and suite names
